+++ database/migrations/2018_10_30_135411_sharing.php
@@ -15,10 +15,10 @@ class Sharing extends Migration
     {
         Schema::dropIfExists('user_album');
         Schema::create('user_album', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('album_id')->unsigned()->index();
+            $table->bigIncrements('id');
+            $table->bigInteger('album_id')->unsigned()->index();
             $table->foreign('album_id')->references('id')->on('albums')->onDelete('cascade');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
